Migrate some more Rust codegen tests to the new framework
