+++ tests/runtime-new/gated-features/runner.rs
@@ -1,0 +1,10 @@
+//@ args = '--features y'
+
+include!(env!("BINDINGS"));
+
+use crate::foo::bar::bindings::{y, z};
+
+fn main() {
+    y();
+    z();
+}
