+++ tests/runtime-new/package-with-version/runner.rs
@@ -1,0 +1,7 @@
+include!(env!("BINDINGS"));
+
+use crate::my::inline::foo::Bar;
+
+fn main() {
+    let _ = Bar::new();
+}
