+++ tests/runtime-new/rust-alternative-bitflags/test.wit
@@ -1,0 +1,19 @@
+package my:inline;
+
+interface t {
+    flags bar {
+        foo,
+        bar,
+        baz
+    }
+
+    get-flag: func() -> bar;
+}
+
+world test {
+  export t;
+}
+
+world runner {
+  import t;
+}
