+++ tests/runtime-new/unused-types/test.rs
@@ -1,0 +1,16 @@
+//@ args = '--generate-unused-types'
+
+include!(env!("BINDINGS"));
+
+use exports::foo::bar::component::Guest;
+use exports::foo::bar::component::UnusedEnum as _;
+use exports::foo::bar::component::UnusedRecord as _;
+use exports::foo::bar::component::UnusedVariant as _;
+
+struct Component;
+
+export!(Component);
+
+impl Guest for Component {
+    fn foo() {}
+}
